% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inkit-package.R
\docType{package}
\name{inkit-package}
\alias{inkit}
\alias{inkit-package}
\title{inkit: Handle and visualise NHP inputs data}
\description{
Handle and visualise data from the inputs data pipeline of the New Hospital Programme (NHP) demand model. Designed primarily to support the NHP Inputs and TPMA explorer web applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/The-Strategy-Unit/inkit}
  \item Report bugs at \url{https://github.com/The-Strategy-Unit/inkit/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{matt.dray@nhs.net}

Authors:
\itemize{
  \item Rhian Davies \email{rhian.davies@nhs.net}
}

Other contributors:
\itemize{
  \item Chris Beeley [contributor]
  \item Fran Barton (\href{https://orcid.org/0000-0002-5650-1176}{ORCID}) [contributor]
  \item Tom Jemmett [contributor]
  \item Ozayr Mohammed [contributor]
  \item Claire Welsh [contributor]
}

}
\keyword{internal}
